Extract accounting close cockpit view model
